Proper program contexts are created, now. Further, basic roles needed to allow users to be assigned roles in program contexts are automatically created. Behave tests to validate the new capability are included.
Also, a minor improvement was made to the alembic wrapper ggrc.migrate
to enable the revision command to create a new version for an extension
module.

Finally, a dependency has been added to requirements.txt for the blink
package to enable signals. We'll use signals as an extension mechanism,
with signals providing an explicit hook for contribution to processing,
such as when a new resource is posted.