Fix script error
This will get rid of the script error. However, with this fix in place,
when a user clicks outside the modal after clicking save, with map and
save checked, depending on when exactly they click, the mapping may be
cancelled. In order to fix this I’ll have to look into how we do, and
how we want to handle clicking outside the modal.