Do not add an owner if already owned
This applies to any model that implements Ownable. The logic used to
add the currently logged in user as an owner regardless of whether or
not it already had an owner. Now it will add the currently logged in
user as an owner ONLY if it does not yet have an owner.