TaskGroup assignee as quick form.
Binding isn’t quite right. The model isn’t updated/saved when modifying
assignee. But otherwise it works fine. I would also like to get a few
pixels of padding at the top of each task group row. Not sure right now
the best way to do that. Using the PBC Request tree.mustache as a
reference makes me think that the spacing is made correct on that
display by the icons. I’m going to ask for some help on this to get it
done correctly.