Map tasks to workflow
This correctly maps tasks to workflow. This also introduces a bug. The
bug is this:
- Click Tasks in the Object Nav
- Click “+ Add Task” on the “Tasks” screen
- Uncaught TypeError: Cannot read property ‘column_view’ of undefined.

I believe that this is happening because the Add Task is trying to look
at the “tasks” mapping on the workflow (it does something using generic
class names in
ggrc/workflows/assets/mustache/tasks/tree_footer.mustache). But it
should be looking at the “direct_tasks” mapping instead.

Will address that bug separately.