The logic in this method should set attributes as we need them. The context_id will be set to the same context_id as the attributable object to which they belong. This will allow RBAC rules set on the attributable object to be used also for the attributes.
This code isn't currently working though. I haven't been able to get the Python code to execute this function. The JSON passed into the PUT call has two elements "custom_attribute_values", which is empty and "custom_attributes", which contains the dictionary using the encoding that Anze specified. The python code in the method in this PR attempts to unpack the dictionary to flush and reset the attribute values from that dictionary. But I haven't been able to configure the SQLAlchemy mappings properly in order to get this method to execute.