Trying to get the buttons to show up on the object page. The logic needs to be: 	If the current user is the contact on the object_review task for the current object, show the buttons. 	Otherwise, if the current user is the object owner or an admin, and the object is not yet submitted 	for review, show "Submit for review" link. Otherwise, show "Review in progress".
I know that the logic isn't quite right in the mustache template. But what I'm working on right now is getting the mustache helper to fire the code defined in the callback function.

I've been told that defer_render is the way to go for this because the mapping to get the object review task must be resolved/refreshed asynchronously.

I'm looking for help on why the callback function isn't firing and how to correct it.