Fix notifications for multiple workflow owners
Since workflows can have any number of owners and not just one, we need
to make sure all of them will get the correct notifications. We also
removed the "by xyz" part in the notifications until it is decided if
that part should display just one owner or list them all.