Remove freezegun from production requirements
Freezegun seems to cause key errors with handling dates even when it's not
used, so we will only have it in the dev enviroment for testing.

This commit will break the dates in show pending notifications, which
will not be relative to the send date anymore. That is still better than
having server errors while trying to send digest notifications.