Fix script error when deleting objects
If you deleted object A that was mapped to object B,
then tried to delete object B without refreshing the page
it caused an error because there was a hanging reference left.

Now get_instance_from_mapping returns undefined when
a hanging reference is encountered instead of erroring out.
This is be easier for the calling code to deal with.