Enable sorting on most fields
Adding attr_sort_field with field value accessor info for the data-field property
enables sorting on that column. There is a lot of repetition and it might be better to
add the ability to default to attr_name, but I felt in this case - because we're
essentially doing config - explicit settings were better than implicit ones.