Make mapping columns work on csv import
Update the mapping column handler to insert mappings and make sure there
are no duplicates. On the dry run, the lines that would be inserted are
also taken into account when generating the "mising object" warnings.
The autogenerated codes that get inserted on imported objects, should
not be allowed in the mapping columns. Those will still generate a
warning even if by the end of an import the code will actually exist.