Add join_query tests
Query parser should work the same if the input is split by or/and and is
then joined with a join_queries function.

The join query function will mostly be used to join generated "relevant"
queries with the user input query string. Generating two query strings
and concatenating them together with and/or statemets would also work as
long as those query strings were in paranthesis. But because you can't
encase full text search in paranthesis, the join_queries function is
used.

The commited tests check if the join_query works with a custom query and
a full text search.