Change priortiy blocks to priority columns
The converter class should handle priority columns instead of blocks.
The reason for this is when some columns will have to be handled in all
blocks before handling other columns. The main case for this are the
mapping columns, but this change is needed for the special mappings such
as audit program and other columns that are direct mappings.