Move get_deep_property into cacheable
This effectively brings back `instance.get_deep_property` but it now works on
pre-parsed property strings which allows for big speedups when plucking deep
properties on a collection of items.

We need get_deep_property to be accessible from outside the tree view controller
in order to use it to fix filtering.