Add prefixes to keys of attribute definitions
We have to add propper prefixes for non standard attributes such as
mappings and custom attributes. Without unique prefixes, custom
attributes with the same name as original attributes would be hidden and
ignored when importing objects. Example would be if we had a custom
attribute with the name "Title". then imports would fail for that
object.