Fix third level tree view rendering
In some cases the `parent_model_name` will be just `Cacheable`. There will be no
definition for it's sub tree, so the filtering will fail.
I added a fallback to parent instance type. And in case even this fail skip
filtering. It's better to ignore filtering and display too much than to throw a
script error and leave the user hanging.