Fix infinite spinner for comments under tasks
This short-circuit failed to trigger post-load events and left the spinner
running indefinitely. I think doing a bit of extra work for the empty `items` is
better than duplicating logic for a proper short circuit so I just removed the
short circuit.

Note: there are still cases of comments not showing up, but this is a separate
issue. This commit only fixes the indefinite spinner.