Make tree view for cycle entries strict
This disables lazy loading in the comments under a cycle task.
The placeholders in there were not styled (at all) and this botched the size
calculations and messed up rendering completely. The end result: displaying only
one comment.
I don't think we need yet another placeholder style since the volume of comments
is rather low so I just disabled lazy loading and this fixes the problem.