Fix data grid with empty code blocks
If there ar no mapped objects found to any previous code block, we have
to ignore that, otherwise the product will not include the lines in the
begining.

It is not possible to have a non empty block after an empty block,
because objects in any consecutive block need to be relevant to any
object in the previous block. So that is why it is safe to just ignore
empty blocks.