Make task_group_id nullable
The sql table structure should match our models. The column
task_group_id can be null in CycleTaskGroup model but not in the
cycle_task_groups table. We need to make the column nullable so that our
database does not fall into some incosistent state where task_group_ids
have values of non existent task groups.

To show the issue create a WF and start a cycle, then delete the TG and
you should see a warning in the console.