Rennder visible placeholders on mouseenter
Sometimes the tree view will glitch and leave a placeholder visible as a "zombie
node". This is especially annoying where there aren't enough elements to trigger
a re-render with a scroll so a page refresh is needed.
This introduces a zombie-slayer: render placeholders on mouseenter. I argue that
hovering over the zombies is expected behaviour and this will make for an
intuitive workaround while we fix the underlying issues in the tree view
rendering.