Make tree view unloading more robust
This drops the `last_visible` variable and runs over the list to see which
element is rendered on ever `draw_visible`. It is implemented efficiently so it
only costs about a microsecond per element in the list. This now correctly
handles the rows rendered by `mouseenter` events and makes the rendering in
general more robust since it tracks less state.