Implement Task start/end export filters
This is a bit special because start and end are several fields whose semantics
depend on the frequency of the associated workflow. I implemented a "macro
expansion" pass over the export filter AST that looks at the right hand side of
the expression (concerning start/end), guesses the frequency semantics based on
the format of the date and rewrites the AST subtree with specific date fields.

Code is ugly and ad-hoc but it does the job so I don't see a point in rewriting
until we need to add more "macro attributes".