Force task dates to be just dates
We must make sure that the start and the end date on the task will be
the same as it will be stored in the database. Adding datetime to start
or end dates would cause the DB to truncate that data and it would break
the valid date range validation.