Fix task date validation
The validats decorator depends on the order of edits. Start date has to
be set before end date. If that's not the case it might miss an error,
which is fine because imports make sure that the order of editing is
correct. This decorator is used mostli for displaying the correct
warning to the user during import of csv files. The real check happens
with schema.CheckConstraint which should work in all cases, but the
error is not user friendly.