Add automatic testing for import errors
User will only see errors that are listed in the dry_run response and
the tests should automatically make sure that those are always the same
as the actual errors that occur on the file import. This commit helps so
that the developer does not have to test dry run and actual import
responses manually.