Allow risks and threat mappings in import/export
This is a quick fix for allowing mapping of risks and threats to all
other models in the app. The propper fix should come later when these
rules are read from each individual module.

Making this fix instead of the propper fix is just for the zucchini
release, and we don't want to change too much code and create any extra
bugs.