Remove zlib from compressed type
Even though we have a blob field to store compressed types, we still get
a warning from the database about some invalid utf8 characters. This is
just to remove that warning.

Performance is not affected much since most of the parameters that are
stored in the compressed type are tiny.