Update selenium test runner
This commit only moves the part of the code that waits for the server to
start befor running selenium tests. This is useful, so that we don't
have to rely on the host machine to expose port 8080. Exposing a single
point would mean that we could only run one instance of the test on a
single host.