Update workflow_state property
This change is to remove the inner for loop in the previous get_state
function which would query the db on each iteration and causing a big
performance hit. The solution here still needs a single query per each
object, which can be improved for collection get.

This PR also adds tests to make sure that the function behaviour is not
changed. The test is written so it can be called as a single test on an
already populated database.