Ignore pylint checks for line length
Line lengths are checked with flake8 and don't need to be checked with
pylint.

The reason why we want to remove this check from pylint is due to urls
in docstrings. Flake8 will not complain if a line is too long and starts
with http:// or https://, whereas pylint will throw a warning. Relying
on flake8 to do line checks is nicer than removing those checks for
pylint in some files.