Clean up workflow export tests
The main part for this commit, is making sure these tests can be run on
a non clean database. Running test_workflow_export worked only because
of the correct test order and if the database was dropped before.
This was relying on the auto-increment id column to start with 1 for
cycles and other objects.

The script 'run_pytests' does not catch it since the database is always
clean and test order never changes. To see this bug you must run the
workflow export tests a couple times in a row.

Changes in this commit include:
- Make tests independent off mysql auto-increment id.
- Fix testing of exported object types. Values should be surrounded with
  commas to avoid false negatives.
- Fix indentation.
- Add basic docstrings.