Remove object_types table
This migration changes polymorphic notifications table, so that it uses
a string for object types instead of foreign keys in the object_types
table. The object_types table was added when we tried to migrate all
polymorphic tables to use foreign keys as an optimization for joins.
This approach was later abandoned for easier to program approach with
strings for object types.