Fix ggrc notifications init
Imports of packages and modules should not execute any code upon
imports. The creation of the Services class was dependent on which
modules have already been loaded when the class was created. This is a
refactor for making sure that ggrc.notifications can be imported
independetly on when the ggrc modules themselves are loaded.