Update custom attributes uniqueness migration
This update contains two fixes. First, it accounts for the fact that
while Python's string comparison is case sensitive, MySQL's is not and
we need to mitigate this by comparing lowercase values. Secondly, even
if a particular CA name is changed, we still need to make sure such a
name is not already in the database.