Fix failure on object save when CAs change
When custom attributes definitions are changed there could be dangling values
left - values with object but with missing definition. At least this is what
happened on the grc-test database. These values are cleaned out when updating an
object but the new logic for checking for mapping CAs tried to load the
definition and resulted in a crash.
This fix simply discards old values of custom attributes on the front end if it
cannot find definitions.