Update pylint diff checker
The check_pylint_diff script used wrong commits while running
'check_pylint_diff develop' on a feature branch. This fix makes sure
that running that on a feature branch works as the help suggests and is
the same as running 'check_pylint_diff' on a merge commit from that
branch.