Use fresh json for logging revisions
There is at least one handler that executes in between an item being added to
cache and log_event happening. This is
`Slugged#ensure_slug_after_flush_postexec`. As a result revision entry stores
a stale slug which looks like corrupted data to the user inside the object
history log. By not using the cached value and recomputing the json we can
avoid this at minimal cost.