Speed up tree view item rendering
Changed the way items are scheduled for rendering so schedule them all
upfront instead of chaining them. This makes them draw faster.

Stopped replacing a rendered item with a placeholder when it is no
longer visible.

Decreased the delay between the user stopping scrolling and the item
rendering starting.

Sped up template rendering of each item by switching open_close
template to use 'renderLive' instead of 'render'. Serializing the
context in the 'render' function was taking about 40% of the rendering
time. It is still taking ~29ms to draw each item.