Fix getting object state for invalid objects
The database can contain some cycle tasks without an end_date. That
causes api calls for any object on that task to fail. The main
difference in the code here is the check if task.end_date is None right
before the date < today check. The rest is just a trivial performance
improvement since we only need to check if any such task exists.

The underlying problem for this will need to be found and solved in a
different PR. The point if this commit is to make sure invalid database
state does not break the app.