Update default_people validator
Default people field on assessment templates is json filed with two
mandatory elements: assessors and verifiers. Both of those can accept
either a valid default people label or a non empty list of user ids. If
the fields contain labels, we should verify them, but for the user list
it's okay to just check if it's not empty, since doing actual id checks
is too expensive.