Improve permissions loading performance
The user object in the session is not eagerly loaded and the
user.object_owners always triggered a new query. The main bottleneck
here was serialization of a lot of objects, not the query itself. This
commit does not change the behaviour of the function but only gets rid
of the object generation overhead.

Benchmarks (on local instance) before and after for 3k owner objects:
- avg  0.11570 - max:  0.18693  - min:  0.08616
- avg  0.03412 - max:  0.10563  - min:  0.01692