Add check if element and header are present
In hard-to-reproduce and quite rare cases (but quite often encountered
when editing files) header is not present when executing \
`ensureElementVisible` on page load, in which case there is an error
thrown:
“cannot access top of undefined” on headerTop = $header.offset().top
call. After that error infopin expansion dies.

This silently ignores that and just skips the execution on page load
(no noticeable issues with that) while still allowing info pin expansion
for later use.