Add a list of valid CA types
Having a class with valid types is useful for autocomplete but not much
use when you have to loop through them or used them as in a function.
This commit just copies the valid types into a list and even those
should eventually be moved to a dict. A dict would be good for
separating what we show to the user and what is stored in the database.