Add exclude option to attribute definitions
Assessment template can have a custom attribute definition mapped to it
without having the actual attribute. Since it is impossible to assign
any value to such an attribute definition we should be able to exclude
it from the object attribute lists. This fixes false missing mandatory
errors on Assessment Template updates via import.