Move deferred wrapper to its own module
The deferred wrapper is used in all our models and mixins and having it
in the mixins.py file causes circular imports which are resolved by
having import statements inside functions. We should avoid that and this
is a step to moving all of our import statements to the top of python
files.