Set modified by instead of modified by id
Both of these options work the same way with how modified_by field is
defined. The benefit of using an actual person object is when we try to
serialize obj, the previous solution made an extra sql query to fetch
the user object from the database. This solution remove one redundant
query per object on collection post requests.