Update benchmark context manager
This commit adds option for the benchmark manager to only output the
summary of requests. If there's enough text to be displayed the output
can affect the benchmark results.

With this modification developers can check which parts of a function
are slow and the overall results only by changing the GGRC_BENCHMARK
environment variable to one of possible options: all, last, only, and
only_last.