Remove redundant permissions check
The is_allowed_create_for is a more specific permission check and can
always be trusted, so the context check above is not needed. But because
the permissions for check flushes the current session, we must make sure
that the object that we are checking permissions for is set up properly.
That is why model posted calls are done above the permissions check (See
how workflow cycles are manually created).

This leaves options that bad listeners can save things even when not
allowed. But generally we should not allow commit calls in the pre
commit listeners.