Add validator for custom attribute value
The custom attribute value can not be validate with normal @validates
decorator because it is dependent on other objects to be set first. This
validator is not run automatically but can be used to help developers
manually validate custom attribute values. Main use case for these
validators should be on POST and PUT requests.