Add update raw function to json builder
The update raw flag is used to indicate which setters are capable of
handling the posted json value on their own. Main use case for this is
custom attribute values where we cant rely on all values existing when
the object is posting. The custom setter for should be able to generate
or update any attribute according to json data.