Fix custom attribute export filters
Since the introduction of object level custom attributes we had to use
attribute names as keys instead of ids to avoid duplicates in the
exported columns. This caused the bug of not having a proper custom
attribute definition id. This fix adds the missing data to all custom
attributes, but the behaviour for object level custom attributes is
random.