Display user's email even if name is not empty
In a column where an email is expected (e.g. all sorts of grid views),
it does not make sense to show an email only if the corresponding
user name exists - sometimes it doesn't, and thus showing an email
is necessary in order to "recognize" a user in a list.