Fix context when cloning audit
When creating a new audit we always have to make sure the context is
set to the program context. This is because of a permission check on
the backend that verifies if the current user can create a new audit in
the context of a program.

We don’t set the program context directly in the clone function, we
only set the parent program. Our audit save function will make sure the
correct context is set.