Remove legacy code
The filter keys and mappings were used for accessing object properties
on the frontend by the Javascript filters. Since these filters are now
done with the query API, these variables aren't used anymore. The code
that used them was removed in one of the previous commits but these were
forgotten.