Fix Relationship setters
This issue was caused by PR #4451. After this PR the person object was
sometimes returning LocalProxy as the source_type when generating
assessments. This caused an error in our memcache invalidation as
LocalProxy is not a valid object type.

I have set the `value.type` as the default value for both
destination_type and source_type instead of value.__class__.__name__,
but because this might cause a regression, I have left this as the
default value.