Add extra assessment CAD title unique check
Assessment global CAD titles must not match any of the assessment
template CAD titles. This makes sure that generated assessments do not
have any name conflicts.
The assessment local CAD must be able to match the assessment template
CAD titles, because those will be created when an assessment is
generated from the assessment template with CAD.