Add CAD backref to Assessment
This will allow us to create an object level custom attribute
definition without the assessment.id. Previously we needed to create an
assessment, do a db.session.flush and then create the custom attribute
definition object and pass in the assessment id.

With this change the db.session.flush will no longer be needed.

Note: Setting the definition this way will only work for Assessments,
this is okay for now since assessment is the only object with object
level custom attribute definitions.