Update object level custom attribute definition generation
The make_transient call and the definition.id = None lines looked like
a hack. This new way is a bit more verbose, but also clearer in what we
are doing.

We are also no longer passing in the assessment.id, because we do not
have it before the db.session.flush instead we pass in the assessment
object. For this to work we will need to add a backref to the
assessment class (see next commit).