Fix adding task group assignee to workflows
The function ensure_assignee_is_workflow_member was broken if more than
one task group was added to workflows in a batch job without flushes
between them. This commit checks if an assignee was already added to
workflow_people on an not flushed workflow.
The change from all() to count() is purely for performance and does not
affect the boolean value that is actually tested.