Only send cycle notifications to people with a wf role
Instead of using the object_person list which contains people that are
mapped without a role we now use the user_roles list because people
without a role on a workflow do not want to receive cycle notifications.

Such people will still receive task notifications if they are assigned
to a task.