Undefer updated_at in guery_helper
_get_last_modified function in services/query_helper was fetching the
updated_at column from the database for every object in the loop. For
the object count query on all objects page on grc-test this resulted in
over 14k queries to the database just to fetch the updated_at column
for every object.

I believe this simple code change will reduce the request time for
fetching object counts from minutes to just a couple of seconds at most.