Fix test object generator
The object generator for testing added a default owners and title fields
to posted data. On objects that do not have these fields such as
Comments and Relationships this caused an error. The second bug that hid
that error was that owners contained invalid data unless the test was
run on a clean database with a person with id 1.

This commit fixes both of the above issues.