Update full text index properties
This commit makes sure that custom attributes in the full text table can
be treated as any normal object attribute.

The property field in full text index did not contain custom attribute
names because in we used to allow duplicate names on objects. Since
duplicate names are now restricted, we do not need to create a unique
property name for all custom attributes.