Add scope clonening handler
Because we need to know which event caused the operation to proceed
we cannot rely on collection_posted but have to instead use
model_posted_after_commit. This splits the cloning operation in two
by first cloning the audit object itself and only when all of that
is done clone the scope of the object.