Fix snapshot model conversion
The content of the snapshot model should never refer to the original
objects. This causes problems when refreshing a snapshot of the object
that was already deleted.

The other fix here is to always use correct links and not mix up ids of
the original objects and snapshots, which will create invalid links.