Organize if statements in mapping handlers
This refactor is to improve readability. If statements check if an
object exists or will exists at the time of import and the else
statement is when there is no object.

We also reuse the objects list for the boolean flags on dry_runs since
a dry run will not do anything with those values.