Fix performance degradation introduced by canjs 2.2
In canjs 2.2 and later the each function includes a call to observe,
which then triggers a huge amount of events on each attr change. These
extra events caused a simple instance.attr(‘title’, ‘new title’) call
to trigger so many events that it increasing the time spent in the
function from ~9ms before canjs 2.2 to 800ms after canjs 2.2.

A similar thing happens in the `_super` function as well so I have
replaced it with simply setting the value. This might cause a few
regressions and we might have to handle a few cases manually, but this
is a good step towards finally upgrading canjs to the latest version.