Fix assessment generation with templates
The validator for custom attribute names needs to have the link between
the assessment and the custom attribute definition to work properly.
This is done by sqlalchemy on flush. In some cases the flush gets called
in the check_permissions step, but this ensures we have all the data
flushed before generating the needed links for assessments.