Fix custom custom attribute revision content
When posting with the old custom attribute API, the custom attribute id
gets set as a string. Since that value is then not read from the
database when log_json is called, the revision content for the id is
also stored as a string. Then the front-end can not find the proper
custom attribute id, and the custom attribute values fail to display.