Fix revision content for object owners
Object owners should be included in our revision content because of
snapshots. Because we first create a request for the object itself and
then for object owners, the create revision of the original object will
not contain information on object owners. That is changed after the
first edit, but if the owners are changed, the old values will be used
in the revision.
This commit makes sure that the correct owners info is stored in the
revision content when creating and editing any object. Downside of this
commit that all objects created via API request will have created and
edited revision, even if it's a brand new object.