Move risks migration to ggrc module
The migration for missing Risk and Threat revisions must be in the ggrc
module so that it runs before the Audit snapshot migration.

The filter for missing objects is added for clean provisioning. When
running the migrations on an empty database, the risks and threats
tables do not exists yet at the time of revisions migration. Those
tables can be safely ignored because if they don't exist they can't
contain any objects with missing revisions.