Join multiple insert statements into one
We want to manually create a single big insert statement for migrations
because the biggest bottleneck is database latency. This is the best way
of reducing the time needed for the migrations.

There is no user defined data in the relationships and all types can
only contain valid model names, thus reducing the danger of a possible
SQL injection enough that we don't have to worry.