Fix import helper tests
The scoped objects have their mappings defined in the snapshot rules and
are not related to allowed mappings, which contains a list of all
possible mappings to a given object. All mappings to scoped objects are
either snapshot mappings or other special mappings, but the mapping
columns must only contain snapshot mappings.