Add read only column for some mapping columns
The mapping to Audits and in audit scope objects is not supported and
those columns are read only. Instead of having a warning for all these
fields, we just add a warring to the column description, so we don't
show too many unneeded warnings on imports that might mask any real
issues.