Update (un)mapping columns set
Audits should be mappable to Snapshots, Assessments and Issues.
Assessments should be mappable to Snapshots, Audits and Issues.
Issues should be mappable to Snapshots, Audits and Assessments.

All non-audit-scope should not be mappable to Audits or Audit-scope.

Audits should not be unmappable from Audit-scope and vice versa.
Snapshots should not be unmappable from Audits and Audit-scope.