Undefer option title column
Since option field should act as a normal attribute on other objects its
value should not be lazily loaded when used with joinedload. This
reduces the number of db queries when using such objects. Since title is
a small filed, and there aren't many different options available, the
performance degradation when such a field is not used, is negligible.