Fix query count tests
Extra permission queries pushed the control query count
to the limit (32/32), but since this is not actually
over I just opted to change the asserLess into
assertLessEqual. This also makes the error message
less confusing as the test will only fail when
the query count is actually OVER the limit.