Edit old Threats migration
Migrations should not use ggrc models and the ones that do should be
disabled after they are run on prod. This fixes issues with creating new
models whose tables do not yet exist.

This change is safe because the migration has already been run on all
deployed instances, and when doing a migration for a fresh database
there can not be any duplicates so the commented line can be ignored.