Add ignore DP attributes on test cleanup
The data platform attribute definitions are currently populated with
the migration and should not be modified added or deleted in any way.
This means we should not clean them for every test. The actual table
attributes is still deleted since that one contains values that are
used and modified by the app.