Add header-based authentication for allowed apps
On every request that requires login,
if X-Appengine-Inbound-Appid is set and is in ALLOWED_QUERYAPI_APP_IDS
and X-ggrc-user is set and contains a JSON object with "email" set,
log in as the user with this email.

X-Appengine-Inbound-Appid is managed (cleaned up and set) by AppEngine
internally and can be trusted.