Fix background tasks for computed attributes
The task id gets set as both header and request parameter when creating
a background task, but for some reason parameter is not received when
calling the computed attributes background task. That is why we should
try to use any id if available.

Please note that this is a fix for a bug that I do not know why it
exits. More investigation is needed on this issue and the this commit
might have to be replaced with a more appropriate solution.