Enable mysql logging by default
This will enable mysql logs by default in docker-compose-clean.yml. The
directory /tmp is used instead of /var/log/mysql in order to avoid any
changes to the container, because mysql user does not have permissions
to write to the /var/log/mysql. The logs themselves are also just used
for development and monitoring, and their persistence is not needed.