Explicitly provide created/updated date for URLs
For old object revisions it can happen that created/updated at dates
for reference URLs are not included in the response. This PR explicitly
adds them to each reference URL object in revision data. On top of
that, it uses Revision object's dates if the latter are not available
in revision content.