Add hacky cav filter
This is a hack to prevent all custom attribute values showing up in all
columns.

The issue behind this is how the mustache helper is used. Instead of
getting what value to display it loops through all values and expect to
receive a value only if the column name matches the cad title. This code
should be removed and the mustache file should only call the helper with
a single value per cell.